Create Smart Saline Bottle System
This file contains code of Smart Saline Bottle system made by us in the Engineering Exploration Project. It checks the level of liquid in the Saline Bottle and messages the nurse or the doctor if the level is low. It also has a buzzer for the nearby nurses to know sonically.